Move host_path, downwardapi_volume and empty_dir into common directory.
